git-svn-id: http://svn.resourcespace.com/svn/rs/branches/20200917_nbraddy_autorotate_change_q10531@16266 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ plugins/image_text/hooks/collection_download.php
@@ -98,8 +98,11 @@ function HookImage_textCollection_downloadModifydownloadfile()
 		{
 		$p=$newdlfile;
 		}
-		
-	unlink($tmpolfile);
+        
+    if(file_exists($tmpolfile))
+        {
+        unlink($tmpolfile);
+        }    
 	
 	return true;
 	}
